rename : useAuth shared/hooks => entities/auth
