+++ client/src/entities/auth/index.ts
@@ -2,3 +2,4 @@ export * as authMutations from './mutations';
 export * as authApi from './apis';
 export * as authType from './type';
 export { default as useAuthStore } from './useAuthStore';
+export { default as useAuth } from './useAuth';
